fix: regoranise handling error response
